# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Kupikupi API
  version: 0.1.0
  description: API draft for the Kupikupi personal shopping agent MVP.
servers:
  - url: https://api.kupikupi.cz/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /watchlists:
    get:
      tags: [Watchlists]
      summary: List watchlists
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: archived
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Watchlists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistList"
    post:
      tags: [Watchlists]
      summary: Create watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistCreate"
      responses:
        "201":
          description: Created watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{watchlist_id}:
    get:
      tags: [Watchlists]
      summary: Get watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Watchlists]
      summary: Replace watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistUpdate"
      responses:
        "200":
          description: Updated watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
    delete:
      tags: [Watchlists]
      summary: Delete watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "204":
          description: Deleted
  /watchlists/{watchlist_id}/pause:
    post:
      tags: [Watchlists]
      summary: Pause watchlist notifications and matching
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "200":
          description: Paused watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{watchlist_id}/archive:
    post:
      tags: [Watchlists]
      summary: Archive watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "200":
          description: Archived watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{watchlist_id}/restore:
    post:
      tags: [Watchlists]
      summary: Restore and activate an archived watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "200":
          description: Restored watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
    WatchlistId:
      name: watchlist_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    WatchlistCreate:
      type: object
      required: [type]
      properties:
        type:
          $ref: "#/components/schemas/WatchlistType"
        product_id:
          type: string
          format: uuid
          nullable: true
        source_request_id:
          type: string
          format: uuid
          nullable: true
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        size_value:
          type: string
          nullable: true
        size_system:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        target_price:
          type: number
          format: decimal
          nullable: true
        target_price_currency:
          type: string
          nullable: true
          example: EUR
        discount_threshold:
          type: number
          format: decimal
          nullable: true
        notify_on_historical_min:
          type: boolean
          default: true
    WatchlistUpdate:
      allOf:
        - $ref: "#/components/schemas/WatchlistCreate"
        - type: object
          properties:
            active:
              type: boolean
            archived:
              type: boolean
    Watchlist:
      allOf:
        - $ref: "#/components/schemas/WatchlistUpdate"
        - type: object
          required: [id, user_id, type, active, archived, created_at]
          properties:
            id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            category:
              type: string
              nullable: true
            brand:
              type: string
              nullable: true
            use_case:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    WatchlistType:
      type: string
      enum: [exact_product, category_rule, brand_rule, agent_request]
    WatchlistList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Watchlist"
        total:
          type: integer
...